Flytter GyldigeDatoer til egen fil.
