+++ src/components/dato/GyldigeDatoer.ts
@@ -1,0 +1,8 @@
+export enum GyldigeDatoer {
+  alle = 'alle',
+  fremtidige = 'fremtidige',
+  tidligere = 'tidligere',
+  tidligereOgSeksMånederFrem = 'tidligereOgSeksMånederFrem',
+  femÅrTidligereOgSeksMånederFrem = 'femÅrTidligereOgSeksMånederFrem',
+  femtiÅrTidligereOgSeksMånederFrem = 'femtiÅrTidligereOgSeksMånederFrem',
+}
