+++ src/components/dato/MånedÅrVelger.tsx
@@ -1,11 +1,11 @@
 import React from 'react';
 import { useSpråkContext } from '../../context/SpråkContext';
 import { hentUid } from '../../utils/autentiseringogvalidering/uuid';
-import { GyldigeDatoer } from './Datovelger';
 import { MonthPicker, useMonthpicker } from '@navikt/ds-react';
 import { hentTekst } from '../../utils/teksthåndtering';
 import { useLokalIntlContext } from '../../context/LokalIntlContext';
 import { hentDatobegrensninger } from '../../utils/dato';
+import { GyldigeDatoer } from './GyldigeDatoer';
 
 interface Props {
   valgtDato: Date | undefined;
