+++ src/søknader/barnetilsyn/steg/6-barnepass/BarnepassSpørsmål.tsx
@@ -17,11 +17,11 @@ import { erÅrsakBarnepassSpmBesvart } from './hjelper';
 import { harValgtSvar } from '../../../../utils/spørsmålogsvar';
 import { EBarnepass, ETypeBarnepassOrdning, IBarnepassOrdning } from '../../models/barnepass';
 import { EPeriode } from '../../../../models/felles/periode';
-import { GyldigeDatoer } from '../../../../components/dato/Datovelger';
 import { erPeriodeGyldigOgInnenforforBegrensning } from '../../../../utils/gyldigeDatoerUtils';
 import { Heading, TextField } from '@navikt/ds-react';
 import { SettDokumentasjonsbehovBarn } from '../../../overgangsstønad/models/søknad';
 import { TittelOgSlettKnapp } from '../../../../components/knapper/TittelOgSlettKnapp';
+import { GyldigeDatoer } from '../../../../components/dato/GyldigeDatoer';
 
 interface Props {
   barn: IBarn;
