+++ src/søknader/felles/steg/1-omdeg/sivilstatus/SøkerHarSøktSeparasjon.tsx
@@ -2,10 +2,11 @@ import React from 'react';
 import FeltGruppe from '../../../../../components/gruppe/FeltGruppe';
 import KomponentGruppe from '../../../../../components/gruppe/KomponentGruppe';
 import AlertStripeDokumentasjon from '../../../../../components/AlertstripeDokumentasjon';
-import { GyldigeDatoer, Datovelger } from '../../../../../components/dato/Datovelger';
+import { Datovelger } from '../../../../../components/dato/Datovelger';
 import { useOmDeg } from '../OmDegContext';
 import { useLokalIntlContext } from '../../../../../context/LokalIntlContext';
 import { hentTekst } from '../../../../../utils/teksthåndtering';
+import { GyldigeDatoer } from '../../../../../components/dato/GyldigeDatoer';
 
 const SøkerHarSøktSeparasjon: React.FC = () => {
   const { sivilstatus, settSivilstatus } = useOmDeg();
