+++ src/søknader/felles/steg/2-bosituasjon/EkteskapsliknendeForhold.tsx
@@ -6,8 +6,9 @@ import FeltGruppe from '../../../../components/gruppe/FeltGruppe';
 import { hentTekst } from '../../../../utils/teksthåndtering';
 import { useLokalIntlContext } from '../../../../context/LokalIntlContext';
 import { harFyltUtSamboerDetaljer } from '../../../../utils/person';
-import { GyldigeDatoer, Datovelger } from '../../../../components/dato/Datovelger';
+import { Datovelger } from '../../../../components/dato/Datovelger';
 import { useBosituasjon } from './BosituasjonContext';
+import { GyldigeDatoer } from '../../../../components/dato/GyldigeDatoer';
 
 export const EkteskapsliknendeForhold: FC = () => {
   const { bosituasjon, settBosituasjon } = useBosituasjon();
