+++ src/søknader/felles/steg/5-aktivitet/utdanning/NårSkalDuElevEllerStudent.tsx
@@ -2,11 +2,11 @@ import React, { useEffect } from 'react';
 import { UnderUtdanning } from '../../../../../models/steg/aktivitet/utdanning';
 import PeriodeDatovelgere from '../../../../../components/dato/PeriodeDatovelger';
 import { tomPeriode } from '../../../../../helpers/tommeSøknadsfelter';
-import { GyldigeDatoer } from '../../../../../components/dato/Datovelger';
 import { hentTekst } from '../../../../../utils/teksthåndtering';
 import { EPeriode } from '../../../../../models/felles/periode';
 import KomponentGruppe from '../../../../../components/gruppe/KomponentGruppe';
 import { useLokalIntlContext } from '../../../../../context/LokalIntlContext';
+import { GyldigeDatoer } from '../../../../../components/dato/GyldigeDatoer';
 
 interface Props {
   utdanning: UnderUtdanning;
