+++ src/søknader/felles/steg/5-aktivitet/utdanning/TarUtdanning.tsx
@@ -26,10 +26,10 @@ import Studiekostnader from './Studiekostnader';
 import { Stønadstype } from '../../../../../models/søknad/stønadstyper';
 import styled from 'styled-components';
 import { erPeriodeGyldigOgInnenforforBegrensning } from '../../../../../utils/gyldigeDatoerUtils';
-import { GyldigeDatoer } from '../../../../../components/dato/Datovelger';
 import { Heading } from '@navikt/ds-react';
 import { hentTekst } from '../../../../../utils/teksthåndtering';
 import { useLokalIntlContext } from '../../../../../context/LokalIntlContext';
+import { GyldigeDatoer } from '../../../../../components/dato/GyldigeDatoer';
 
 const LesMerTekstUnderSidetittel = styled(LesMerTekst)`
   margin-top: -2rem;
