+++ src/utils/gyldigeDatoerUtils.ts
@@ -1,7 +1,7 @@
-import { GyldigeDatoer } from '../components/dato/Datovelger';
 import { addMonths, addYears, compareAsc, isEqual, subYears } from 'date-fns';
 import { dagensDato, erGyldigDato, strengTilDato } from './dato';
 import { IPeriode } from '../models/felles/periode';
+import { GyldigeDatoer } from '../components/dato/GyldigeDatoer';
 
 // Brukes for å ikke vise nesteknapp vis dato er ugyldig format eller utenfor begrensninger
 export const erDatoGyldigOgInnenforBegrensning = (
